requirements.txt
fastapi 
uvicorn 
pyarrow
sqlalchemy 
databases 
python-jose
cryptography
faker
pandas
requests
python-multipart